Revert "LPS-72208 Move configuration interface for Asset Publisher to another package and fix imports. No logical changes"
This reverts commit 2f272e85ee5be3aeed459bcc58410ada0f1c3db5.
